Remove PromExploreQueryEditor and PromQueryEditor
