+++ public/app/plugins/datasource/prometheus/components/PromQueryEditorByApp.tsx
@@ -1,12 +1,9 @@
 import React, { memo } from 'react';
 
 import { CoreApp } from '@grafana/data';
-import { config } from '@grafana/runtime';
 
 import { PromQueryEditorSelector } from '../querybuilder/components/PromQueryEditorSelector';
 
-import { PromExploreQueryEditor } from './PromExploreQueryEditor';
-import { PromQueryEditor } from './PromQueryEditor';
 import { PromQueryEditorForAlerting } from './PromQueryEditorForAlerting';
 import { PromQueryEditorProps } from './types';
 
@@ -16,16 +13,8 @@ export function PromQueryEditorByApp(props: PromQueryEditorProps) {
   switch (app) {
     case CoreApp.CloudAlerting:
       return <PromQueryEditorForAlerting {...props} />;
-    case CoreApp.Explore:
-      if (config.featureToggles.promQueryBuilder) {
-        return <PromQueryEditorSelector {...props} />;
-      }
-      return <PromExploreQueryEditor {...props} />;
     default:
-      if (config.featureToggles.promQueryBuilder) {
-        return <PromQueryEditorSelector {...props} />;
-      }
-      return <PromQueryEditor {...props} />;
+      return <PromQueryEditorSelector {...props} />;
   }
 }
