+++ public/app/plugins/datasource/prometheus/querybuilder/components/PromQueryCodeEditor.tsx
@@ -4,7 +4,6 @@ import React from 'react';
 import { GrafanaTheme2 } from '@grafana/data';
 import { useStyles2 } from '@grafana/ui';
 
-import { testIds } from '../../components/PromQueryEditor';
 import PromQueryField from '../../components/PromQueryField';
 import { PromQueryEditorProps } from '../../components/types';
 
@@ -28,7 +27,6 @@ export function PromQueryCodeEditor(props: Props) {
         onChange={onChange}
         history={[]}
         data={data}
-        data-testid={testIds.editor}
         app={app}
       />
